explorer/service/pyproject.toml
[project]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.92.0",
    "uvicorn[standard]>=0.20.0",
    "pysimdjson>=5.0.2",
    "marisa-trie>=0.7.8",
]

[tool.pdm]

